Add available queries to category, add map (query, price) to announcement

+++ chaincode/query.go
@@ -8,13 +8,13 @@ import (
 
 // Query object that represents a query in the World State
 type Query struct {
-	QueryId        string    `json:"queryId"`
-	AnnouncementId string    `json:"announcementId"`
-	IssuerId       string    `json:"issuerId"`
-	Query          string    `json:"query"`
-	Price		   int		 `json:"price"`
-	Response       string    `json:"Response"`
-	InsertedAt     time.Time `json:"insertedAt"`
+	QueryId        string     `json:"queryId"`
+	AnnouncementId string     `json:"announcementId"`
+	IssuerId       string     `json:"issuerId"`
+	Query          string     `json:"query"`
+	Price		   float32	  `json:"price"`
+	Response       string     `json:"Response"`
+	InsertedAt     time.Time  `json:"insertedAt"`
 }
 
 // Serialize formats the Query as JSON bytes
